+++ inst/tinytest/test_describe.R
@@ -1,0 +1,5 @@
+## Read fake data
+artefacts <- read.csv(system.file("tinytest/fake.csv", package = "arkhe"))
+
+# Describe =====================================================================
+expect_stdout(describe(artefacts))